hw_2/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic[email]==2.4.2
sqlalchemy==2.0.20
pydantic-settings==2.0.3
aiosqlite==0.19.0
passlib==1.7.4
bcrypt==4.0.0
python-jose==3.3.0
pytest
httpx==0.25.0
trio